video-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiomysql==0.2.0
python-dotenv==1.0.1
gunicorn==22.0.0
